demo/requirements.txt
python-dateutil
Django<1.8
django-extensions
Sphinx
sphinx_rtd_theme